refactor: 이미지 api apis/image 폴더로 이동
